Scoreboard

+++ ETU/Game/Player.h
@@ -11,8 +11,8 @@ public:
 public:
 	static const float PLAYER_MOVE_SPEED;
 	static const int INITIAL_LIFE_COUNT;
-	static const int MAX_IMMUNE_TIME;
-	static const int MAX_BONUS_TIME;
+	static const float MAX_IMMUNE_TIME;
+	static const float MAX_BONUS_TIME;
 	Player();
 	virtual bool init(const Level01ContentManager& contentManager);
 	bool isGunBonusActive();
